pyproject.toml
[project]
name = "equipment-availablity-service"
version = "0.0.1"
dependencies = [
    "beautifulsoup4",
    "flask",
    "jsonschema",
    "lxml",
    "python-dateutil",
    "python-dotenv",
    "python-json-logger",
    "PyYAML",
    "requests",
    "waitress",
]
[project.optional-dependencies]
test = [
    "debugpy",
    "pipdeptree",
    "pytest",
    "pytest-cov",
    "requests-mock",
    "ruff",
    "pycodestyle"
]
[project.scripts]
equipment-availability-service = "equipment_availability.server:run"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
# preview is required to enable many pycodestyle rules see
# https://docs.astral.sh/ruff/rules/
preview = true

# Enable pycodestyle errors and warnings
select = ["E", "W"]

